Update reward_function.py
Version 6. Cloned V3 using a modified version of V4-2's code. We wanted to implement higher reward initializations (i.e., 10.0, 5.0, 1.0, 1e-3) instead of (1.0, 0.5, 0.1, and 1e-6). We also wanted to try adding and subtracting rewards instead of using multiplication to increase or decrease rewards by a percentage. We will also have a check at the end of the reward function to see if the reward is negative, and if it is we will change the reward to a small value such as 1e-6. We wanted to start these changes from a model that had a successful Reward graph such as V3.